SAK-6115 - Improve JSR-168 Portlets
Built a brand new variant of the Portlet called SakaiLaunch (added to SakaiGallery and SakaiTree).  SakaiLaunch 
is the best of the portlets and probably makes all of the others obsolete.  

SakaiLaunch eliminates all but one iframe and provides a very very breadcrumb navigation.  It 
also autosizes like the Charon portal so there are *no* scrollbars.  There is only one tool 
on the screen at one time.  The reset button is folded into navigation.

Improved the debugging of the portlets - added a property to control debug printing.  I left the portlets 
chatty by default to help in debugging.

Figured out how to deploy the portlets in Pluto 1.1 and tested the portlets both in Pluto 1.0 and Pluto 1.1.

Note - If you want to use Pluto 1.1, you should look at this blog post.  

http://www.dr-chuck.com/csev-blog/000194.html

Related JIRAs: SAK-2473, SAK-6113, SAK-3134, SAK-5827
Related Commits: 14062, 14063, 14064, 14065




git-svn-id: https://source.sakaiproject.org/svn/portlet/trunk@14066 66ffb92e-73f9-0310-93c1-f5514f145a0a